Modify Transpiler to work with gpu
