+++ src/gpu/lib.ts
@@ -1,20 +1,6 @@
 import { GPU } from 'gpu.js'
 
-export function __createKernel1() {
-  const gpu = new GPU()
-  const multiplyMatrix = gpu
-    .createKernel(function(a, b) {
-      let sum = 0
-      for (let i = 0; i < 512; i++) {
-        sum += a[this.thread.y as any][i] * b[i][this.thread.x]
-      }
-      return sum
-    }, {})
-    .setOutput([512, 512])
-
-  return multiplyMatrix([1, 2, 3], [1, 2, 3])
-}
-
+// helper function to build 2D array output
 function build2DArray(arr: Float32Array[][], end: any): Float32Array[][] {
   const res = []
   for (let i = 0; i < end[0]; i++) {
@@ -23,6 +9,7 @@ function build2DArray(arr: Float32Array[][], end: any): Float32Array[][] {
   return res
 }
 
+// helper function to build 3D array output
 function build3DArray(arr: Float32Array[][][], end: any): Float32Array[][][] {
   const res = []
   for (let i = 0; i < end[0]; i++) {
@@ -35,35 +22,20 @@ function build3DArray(arr: Float32Array[][][], end: any): Float32Array[][][] {
   return res
 }
 
+/* main function that runs code on the GPU (using gpu.js library)
+ * @end : end bounds for array
+ * @extern : external variable definitions {}
+ * @f : function run as on GPU threads
+ */
 export function __createKernel(end: any, extern: any, f: any) {
   const gpu = new GPU()
+
+  // external variables to be in the GPU
   const out = { constants: {} }
   out.constants = extern
 
-  console.log(f)
-
-  // function ff(a : any, b : any) {
-  //     return a + b;
-  // }
-
-  // // const r2 = function() {
-  // //     let r = 0;
-  // //     for (let j = 0; j < 5; j = j+1) {
-
-  // //     }
-  // //     return r + this.constants.y
-  // // }
-
-  // gpu.addFunction(ff)
-
-  const g = gpu.createKernel(f, out).setOutput(end)
-
-  console.log(end)
-  console.log(out)
-
-  const res = g() as any
-  console.log('HIHI')
-  console.log(res)
+  const gpuFunction = gpu.createKernel(f, out).setOutput(end)
+  const res = gpuFunction() as any
 
   if (end.length === 1) return Array.from(res)
   else if (end.length === 2) return build2DArray(res, end)
