+++ src/gpu/transfomer.ts
@@ -1,42 +1,44 @@
 import * as es from 'estree'
 import { ancestor, simple, make } from 'acorn-walk/dist/walk'
 import * as create from '../utils/astCreator'
-import GPULoopDetecter from './loopChecker'
-// import { parse } from 'acorn'
-import { generate } from 'astring'
+import GPULoopVerifier from './loopVerifier'
+import GPUBodyVerifier from './bodyVerifier'
 
-/*x
-  takes in a node and decides whether it is possible to break down into GPU     library code
-*/
+/*
+ * GPU Transformer runs through the program and transpiles for loops to GPU code
+ * Upon termination, the AST would be mutated accordingly
+ * e.g.
+ * let res = 0;
+ * for (let i = 0; i < 5; i = i + 1) {
+ *    res[i] = 5;
+ * }
+ * would become:
+ * let res = 0;
+ * res = __createKernel([5], {}, function() { return 5; })
+ */
 class GPUTransformer {
+  // program to mutate
   program: es.Program
+
+  // helps reference the main function
   static globalIds = {
     __createKernel: create.identifier('__createKernel')
   }
 
-  // functions needed
-  inputArray: es.Expression
   outputArray: es.Identifier
-  updateFunction: any
-
-  // parallelizable body
   innerBody: any
-
-  // GPU Loops
   counters: string[]
   end: es.Expression[]
-
   state: number
-
   localVar: Set<string>
   outerVariables: any
-
   targetBody: any
 
   constructor(program: es.Program) {
     this.program = program
   }
 
+  // transforms away top-level for loops if possible
   transform = () => {
     const gpuTranspile = this.gpuTranspile
 
@@ -53,12 +55,24 @@ class GPUTransformer {
     // tslint:enable
   }
 
-  // given a for loop, this function returns a GPU specific body
+  /*
+   * Here we transpile away a for loop:
+   * 1. Check if it meets our specifications
+   * 2. Get external variables + target body (body to be run across gpu threads)
+   * 3. Build a AST Node for (2) - this will be given to (8)
+   * 4. In body, update all math_* calls to become Math.* calls
+   * 5. In body, update all external variable references
+   * 6. In body, update reference to counters
+   * 7. Change assignment in body to a return statement
+   * 8. Call __createKernel and assign it to our external variable
+   */
   gpuTranspile = (node: es.ForStatement) => {
-    // set state
+    // initialize our class variables
     this.state = 0
     this.counters = []
     this.end = []
+
+    // 1. verification of outer loops + body
 
     this.checkOuterLoops(node)
 
@@ -67,46 +81,29 @@ class GPUTransformer {
       return
     }
 
-    console.log(this.counters)
-    console.log(this.end)
-
-    this.checkBody(this.innerBody)
-    if (this.state === 0) {
+    const verifier = new GPUBodyVerifier(this.innerBody, this.counters)
+    if (verifier.state === 0) {
       return
     }
 
-    // we know for loop is parallelizable, so let's transpile
+    this.state = verifier.state
+    this.outputArray = verifier.outputArray
+    this.localVar = verifier.localVar
 
+    // 2. get external variables + the main body
     this.getOuterVariables()
-
-    // get proper body using state
     this.getTargetBody(node)
 
-    // get a dictionary as global variables
-    console.log(this.outerVariables)
-    const p: es.Property[] = []
+    // 3. Build a AST Node of all outer variables
+    const externObject: es.Property[] = []
     for (const key in this.outerVariables) {
       if (this.outerVariables.hasOwnProperty(key)) {
         const val = this.outerVariables[key]
-        p.push(create.property(key, val))
+        externObject.push(create.property(key, val))
       }
     }
 
-    console.log(generate(this.targetBody))
-    console.log('MUTATING')
-
-    // create function?
-    const params: es.Identifier[] = []
-    for (let i = 0; i < this.state; i++) {
-      params.push(create.identifier(this.counters[i]))
-    }
-
-    const counters: string[] = []
-    for (let i = 0; i < this.state; i = i + 1) {
-      counters.push(this.counters[i])
-    }
-
-    // change math functions to Math.__
+    // 4. Update all math_* calls to become Math.*
     simple(this.targetBody, {
       CallExpression(nx: es.CallExpression) {
         if (nx.callee.type !== 'Identifier') {
@@ -114,9 +111,7 @@ class GPUTransformer {
         }
 
         const functionName = nx.callee.name
-        console.log(functionName)
         const term = functionName.split('_')[1]
-        console.log(term)
         const args: es.Expression[] = nx.arguments as any
 
         create.mutateToCallExpression(
@@ -127,16 +122,19 @@ class GPUTransformer {
       }
     })
 
-    // change global variables to be a member access
+    // 5. Update all external variable references in body
+    // e.g. let res = 1 + y; where y is an external variable
+    // becomes let res = 1 + this.constants.y;
+
     const names = [this.outputArray.name, ...this.counters, 'Math']
     const locals = this.localVar
     simple(this.targetBody, {
       Identifier(nx: es.Identifier) {
+        // ignore these names
         if (names.includes(nx.name) || locals.has(nx.name)) {
           return
         }
 
-        console.log(nx)
         create.mutateToMemberExpression(
           nx,
           create.memberExpression(create.identifier('this'), 'constants'),
@@ -145,10 +143,19 @@ class GPUTransformer {
       }
     })
 
-    // change any counters to member access
+    // 6. Update reference to counters
+    // e.g. let res = 1 + i; where i is a counter
+    // becomes let res = 1 + this.thread.x;
+
+    // depending on state the mappings will change
     let threads = ['x']
     if (this.state === 2) threads = ['y', 'x']
     if (this.state === 3) threads = ['z', 'y', 'x']
+
+    const counters: string[] = []
+    for (let i = 0; i < this.state; i = i + 1) {
+      counters.push(this.counters[i])
+    }
 
     simple(this.targetBody, {
       Identifier(nx: es.Identifier) {
@@ -173,7 +180,8 @@ class GPUTransformer {
       }
     })
 
-    // change assignment to a return
+    // 7. Change assignment in body to a return statement
+
     ancestor(this.targetBody, {
       AssignmentExpression(nx: es.AssignmentExpression, ancstor: es.Node[]) {
         // assigning to local val, it's okay
@@ -185,54 +193,32 @@ class GPUTransformer {
           return
         }
 
-        console.log(nx.right.type)
-        console.log(nx.right)
-
         const sz = ancstor.length
-        console.log(sz)
-        console.log(ancstor[sz - 1])
-        console.log(ancstor[sz - 2])
-
         create.mutateToReturnStatement(ancstor[sz - 2], nx.right)
       }
     })
 
-    const f = create.functionExpression([], this.targetBody)
-    console.log(generate(f))
-
-    // mutate node to assignment expression
+    // 8. we transpile the loop to a function call, __createKernel
+    const kernelFunction = create.functionExpression([], this.targetBody)
     create.mutateToExpressionStatement(
       node,
       create.assignmentExpression(
         this.outputArray,
         create.callExpression(
           GPUTransformer.globalIds.__createKernel,
-          [create.arrayExpression(this.end), create.objectExpression(p), f],
+          [create.arrayExpression(this.end), create.objectExpression(externObject), kernelFunction],
           node.loc!
         )
       )
     )
   }
 
-  // get the body that we want to parallelize
-  getTargetBody(node: es.ForStatement) {
-    let mv = this.state
-    this.targetBody = node
-    while (mv > 1) {
-      this.targetBody = this.targetBody.body.body[0]
-      mv--
-    }
-    this.targetBody = this.targetBody.body
-  }
-
-  /** checker our special loops */
-
-  // count all the for loops
+  // verification of outer loops using our verifier
   checkOuterLoops = (node: es.ForStatement) => {
     let currForLoop = node
     while (currForLoop.type === 'ForStatement') {
       this.innerBody = currForLoop.body
-      const detector = new GPULoopDetecter(currForLoop)
+      const detector = new GPULoopVerifier(currForLoop)
 
       if (!detector.ok) {
         break
@@ -253,133 +239,37 @@ class GPUTransformer {
     }
   }
 
-  /** GPU BODY CHECK */
-
-  // get property access identifiers for the result statement
-  getPropertyAccess = (node: es.MemberExpression): string[] => {
-    const res: string[] = []
-    let ok: boolean = true
-
-    let curr: any = node
-    while (curr.type === 'MemberExpression') {
-      if (curr.property.type !== 'Identifier') {
-        ok = false
-        break
-      }
-
-      res.push(curr.property.name)
-      curr = curr.object
+  /*
+   * Based on state, gets the correct body to me run across threads
+   * e.g. state = 2 (2 top level loops skipped)
+   * for (...) {
+   *    for (...) {
+   *      let x = 1;
+   *      res[i] = x + 1
+   *    }
+   * }
+   *
+   * returns:
+   *
+   * {
+   *  let x = 1;
+   *  res[i] = x + 1
+   * }
+   */
+  getTargetBody(node: es.ForStatement) {
+    let mv = this.state
+    this.targetBody = node
+    while (mv > 1) {
+      this.targetBody = this.targetBody.body.body[0]
+      mv--
     }
-
-    if (!ok) {
-      return []
-    }
-
-    this.outputArray = curr
-    return res.reverse()
-  }
-
-  // check gpu parallelizable body
-  checkBody = (node: es.Statement) => {
-    let ok: boolean = true
-
-    // check illegal statements
-    simple(node, {
-      FunctionDeclaration() {
-        ok = false
-      },
-      ArrowFunctionExpression() {
-        ok = false
-      },
-      ReturnStatement() {
-        ok = false
-      },
-      BreakStatement() {
-        ok = false
-      },
-      ContinueStatement() {
-        ok = false
-      }
-    })
-
-    if (!ok) {
-      return
-    }
-
-    // check function calls are only to math_*
-    const mathFuncCheck = new RegExp(/^math_[a-z]+$/)
-    simple(node, {
-      CallExpression(nx: es.CallExpression) {
-        if (nx.callee.type !== 'Identifier') {
-          ok = false
-          return
-        }
-
-        const functionName = nx.callee.name
-        if (!mathFuncCheck.test(functionName)) {
-          ok = false
-          return
-        }
-      }
-    })
-
-    if (!ok) {
-      return
-    }
-
-    // get all local variables
-    const localVar = new Set<string>()
-    simple(node, {
-      VariableDeclaration(nx: es.VariableDeclaration) {
-        if (nx.declarations[0].id.type === 'Identifier') {
-          localVar.add(nx.declarations[0].id.name)
-        }
-      }
-    })
-
-    this.localVar = localVar
-
-    // check all assignments and make sure only one global res var assignment
-    const resultExpr: es.AssignmentExpression[] = []
-    simple(node, {
-      AssignmentExpression(nx: es.AssignmentExpression) {
-        // assigning to local val, it's okay
-        if (nx.left.type === 'Identifier' && localVar.has(nx.left.name)) {
-          return
-        }
-
-        resultExpr.push(nx)
-      }
-    })
-
-    // too many assignments!
-    if (resultExpr.length !== 1) {
-      return
-    }
-
-    // not assigning to array
-    if (resultExpr[0].left.type !== 'MemberExpression') {
-      return
-    }
-
-    // check res assignment and its counters
-    const res = this.getPropertyAccess(resultExpr[0].left)
-    if (res.length === 0 || res.length > this.counters.length) {
-      return
-    }
-
-    for (let i = 0; i < this.counters.length; i++) {
-      if (res[i] !== this.counters[i]) break
-      this.state++
-    }
-
-    if (this.state > 3) this.state = 3
+    this.targetBody = this.targetBody.body
   }
 
   // get all variables defined outside the block (on right hand side)
   getOuterVariables() {
     // set some local variables for walking
-    const curr: es.BlockStatement = this.innerBody
+    const curr = this.innerBody
     const localVar = this.localVar
     const counters = this.counters
     const output = this.outputArray.name
